+++ chewy/src/main/java/com/chewy/fwd/dao/MemberDao.java
@@ -5,6 +5,12 @@ import java.util.List;
 import com.chewy.fwd.vo.MemberVo;
 
 public interface MemberDao {
+	
+	// 회원가입 페이지
+		public void register(MemberVo memberVo) throws Exception;
+		
+		// 회원가입 인서트
+		public void insertRegister(MemberVo memberVo) throws Exception;
 
 	public List<MemberVo> selectEmail(MemberVo memberVo) throws Exception;
