+++ chewy/src/main/java/com/chewy/fwd/dao/MemberDaoImpl.java
@@ -15,6 +15,19 @@ public class MemberDaoImpl implements MemberDao{
 	@Autowired
 	private SqlSession sqlSession;
 	
+	// 회원가입 페이지
+		@Override
+		public void register(MemberVo memberVo) throws Exception{
+			
+		}
+		
+		// 회원가입 인서트
+		 @Override
+		 public void insertRegister(MemberVo memberVo) throws Exception{
+			 MemberMapper mapper = sqlSession.getMapper(MemberMapper.class);
+			 mapper.insertRegister(memberVo);
+		 }
+	
 	@Override
 	public List<MemberVo> selectEmail(MemberVo memberVo) throws Exception {
 		MemberMapper mapper = sqlSession.getMapper(MemberMapper.class);
